+++ app/controllers/api/v1/widget/direct_uploads_controller.rb
@@ -1,4 +1,5 @@
 class Api::V1::Widget::DirectUploadsController < ActiveStorage::DirectUploadsController
+  include WebsiteTokenHelper
   before_action :set_web_widget
   before_action :set_contact
 
@@ -17,33 +18,8 @@ class Api::V1::Widget::DirectUploadsController < ActiveStorage::DirectUploadsCon
 
   def direct_upload_json(blob)
     blob.as_json(root: false, methods: :signed_id).merge(direct_upload: {
-      url: blob.service_url_for_direct_upload,
-      headers: blob.service_headers_for_direct_upload
-    })
-  end
-
-  def permitted_params
-    params.permit(:website_token)
-  end
-
-  def auth_token_params
-    @auth_token_params ||= ::Widget::TokenService.new(token: request.headers['X-Auth-Token']).decode_token
-  end
-
-  def set_web_widget
-    @web_widget = ::Channel::WebWidget.find_by!(website_token: permitted_params[:website_token])
-    @current_account = @web_widget.account
-  end
-
-  def set_contact
-    @contact_inbox = @web_widget.inbox.contact_inboxes.find_by(
-      source_id: auth_token_params[:source_id]
-    )
-    @contact = @contact_inbox&.contact
-    raise ActiveRecord::RecordNotFound unless @contact
-  end
-
-  def permitted_params
-    params.permit(:website_token)
+                                                           url: blob.service_url_for_direct_upload,
+                                                           headers: blob.service_headers_for_direct_upload
+                                                         })
   end
 end
