+++ app/javascript/widget/components/ChatAttachment.vue
@@ -27,7 +27,6 @@ import { BUS_EVENTS } from 'shared/constants/busEvents';
 import FluentIcon from 'shared/components/FluentIcon/Index.vue';
 import { DirectUpload } from 'activestorage';
 import { mapGetters } from 'vuex';
-import { setHeader } from 'widget/helpers/axios';
 
 export default {
   components: { FluentIcon, FileUpload, Spinner },
@@ -54,7 +53,7 @@ export default {
       return fileType.includes('image') ? 'image' : 'file';
     },
     async onFileUpload(file) {
-      if (true) {
+      if (this.globalConfig.directUploadsEnabled) {
         this.onDirectFileUpload(file);
       } else {
         this.onIndirectFileUpload(file);
@@ -70,11 +69,13 @@ export default {
           const { websiteToken } = window.chatwootWebChannel;
           const upload = new DirectUpload(
             file.file,
-            `/api/v1/widget/direct_uploads?website_token=${websiteToken}`, {
-            directUploadWillCreateBlobWithXHR: xhr => {
-              xhr.setRequestHeader('X-Auth-Token', window.authToken);
+            `/api/v1/widget/direct_uploads?website_token=${websiteToken}`,
+            {
+              directUploadWillCreateBlobWithXHR: xhr => {
+                xhr.setRequestHeader('X-Auth-Token', window.authToken);
+              },
             }
-          });
+          );
 
           upload.create((error, blob) => {
             if (error) {
